requirements.txt
Flask
Flask-SQLAlchemy
pandas
scikit-learn
xgboost
shap
python-dotenv
gunicorn
google-generativeai
flask-login
werkzeug
